requirements.txt
boto3>=1.4.3
bs4
cassandra-driver>=3.12.0
dnspython
dogpile.cache
eventlog-writer==0.4.5
flask-restful-swagger==0.19
Flask-RESTful==0.3.5
Flask==0.10.1
functional
ipaddress
Jinja2
jsonpath-rw
MarkupSafe
numpy
opentracing==1.2.2
opentracing_utils
prometheus_client
psycopg2
pykube
pymemcache>=1.4.3
pymongo==3.0.3
pymssql>=2.1.3
PyMySQL==0.6.1
pysnmp
python-ldap==2.4.19
python-snappy
pythonwhois==2.4.3
pytz
PyYAML
redis
requests>=2.7.0
setproctitle
stups-tokens>=1.0.14
subprocess32
timeperiod2==0.1
Yapsy>=1.10.423

